requirements.txt
redis
APScheduler
sanic
kafka-python
PyYAML
SQLAlchemy
stix2
